Revert "disable tests for now"
This reverts commit 559c7b0972fd0923c11eaf49790f34e70d6fb6ed.

+++ contracts/Vault/libraries/LibKeepHelpers.sol
@@ -1,7 +1,11 @@
 pragma solidity 0.8.4;
 
 library LibKeepHelpers {
-    function findAddIndex(address _item, address[] memory addressArray) internal pure returns (uint256 i) {
+    function findAddIndex(address _item, address[] memory addressArray)
+        internal
+        pure
+        returns (uint256 i)
+    {
         for (i; i < addressArray.length; i++) {
             //using the conventional method since we cannot have duplicate addresses
             if (addressArray[i] == _item) {
@@ -10,7 +14,11 @@ library LibKeepHelpers {
         }
     }
 
-    function findUintIndex(uint256 _item, uint256[] memory noArray) internal pure returns (uint256 i) {
+    function findUintIndex(uint _item, uint[] memory noArray)
+        internal
+        pure
+        returns (uint256 i)
+    {
         for (i; i < noArray.length; i++) {
             if (noArray[i] == _item) {
                 return i;
@@ -18,7 +26,7 @@ library LibKeepHelpers {
         }
     }
 
-    function removeUint(uint256[] storage _noArray, uint256 to) internal {
+    function removeUint(uint[] storage _noArray, uint to) internal {
         require(_noArray.length > 0, "Non-elemented number array");
         uint256 index = findUintIndex(to, _noArray);
         if (_noArray.length == 1) {
@@ -47,8 +55,8 @@ library LibKeepHelpers {
         }
     }
 
-    function _inUintArray(uint256[] memory _array, uint256 _targ) internal pure returns (bool exists_) {
-        if (_array.length > 0) {
+    function _inUintArray(uint256[] memory _array,uint256 _targ) internal pure returns (bool exists_) {
+      if(_array.length>0){
             for (uint256 i; i < _array.length; i++) {
                 if (_targ == _array[i]) {
                     exists_ = true;
@@ -57,8 +65,8 @@ library LibKeepHelpers {
         }
     }
 
-    function _inAddressArray(address[] memory _array, address _targ) internal pure returns (bool exists_) {
-        if (_array.length > 0) {
+    function _inAddressArray(address[] memory _array,address _targ) internal pure returns (bool exists_) {
+      if(_array.length>0){
             for (uint256 i; i < _array.length; i++) {
                 if (_targ == _array[i]) {
                     exists_ = true;
@@ -66,6 +74,4 @@ library LibKeepHelpers {
             }
         }
     }
-
-
 }
